[Fix][kubectl-plugin]: make `version` handle digests (#2876)
Right now `kubectl ray version` returns only the image digest if the KubeRay
operator uses an image reference that contains both tag and digest. This change
returns both and also handles the case where only the digest is present.

We depend on the https://github.com/novln/docker-parser library to simplify
image reference parsing logic except in the case where both tag and digest are used
which the library doesn't seem to handle. See [this issue][1].

[1]: https://github.com/novln/docker-parser/issues/3